Fix handling of primitive types in Java (#661)
We adapted the live test for primitive types to Java. Running it
revealed five bugs in the Java generator:

- The XML deserializer called the non-existent ``readContentAsInt``
  and ``readContentAsFloat`` methods.
- The JSON serializer called ``toJsonValue`` instead of the generated
  ``toJsonNode``.
- ``tryLongFrom`` rejected JSON integers that Jackson had parsed as
  ``IntNode`` instead of ``LongNode``.
- The XML serializer omitted the start and end element tags for
  non-optional byte arrays.
- The generated copying test called ``.get()`` on a non-Optional byte
  array getter.

We fix all five issues, and the live tests now pass.

+++ aas_core_codegen/java/transpilation.py
@@ -41,7 +41,7 @@ from aas_core_codegen.parse import tree as parse_tree
 PRIMITIVE_TYPE_MAP = {
     intermediate_type_inference.PrimitiveType.BOOL: Stripped("Boolean"),
     intermediate_type_inference.PrimitiveType.INT: Stripped("Long"),
-    intermediate_type_inference.PrimitiveType.FLOAT: Stripped("Float"),
+    intermediate_type_inference.PrimitiveType.FLOAT: Stripped("Double"),
     intermediate_type_inference.PrimitiveType.STR: Stripped("String"),
     intermediate_type_inference.PrimitiveType.BYTEARRAY: Stripped("byte[]"),
 }
